--- .github/workflows/pytorch_tests.yml
@@ -1,199 +1,0 @@
-name: PyTorch Tests
-
-on:
-  push:
-    branches:
-      - main
-  pull_request:
-    types: [opened, reopened, synchronize, ready_for_review]
-    paths-ignore:
-      - "examples/**"
-      - "llvm/**"
-      - "mlir/**"
-      - "targets/et/**"
-      - "targets/tenstorrent/**"
-      - "tutorial/**"
-  schedule:
-    - cron: "0 4 * * *"
-
-jobs:
-
-  pytorch-tests-openmp:
-    runs-on:
-      group: dahlia
-      labels: Linux
-    container:
-      image: daisytuner/docc-run-env-llvm19-ubuntu-24.04:latest-amd64
-      options: >-
-        --cap-add=PERFMON
-    strategy:
-      fail-fast: false
-      matrix:
-        python-version: ["3.11", "3.12", "3.13", "3.14"]
-
-    steps:
-      - uses: actions/checkout@v4
-        with:
-          submodules: recursive
-
-      - name: Set up Python ${{ matrix.python-version }}
-        uses: actions/setup-python@v5
-        with:
-          python-version: ${{ matrix.python-version }}
-
-      - name: Create virtual environment
-        run: |
-          python -m venv .venv
-          echo "$PWD/.venv/bin" >> $GITHUB_PATH
-
-      - name: Mark GitHub Actions workdir as safe
-        run: git config --global --add safe.directory "$GITHUB_WORKSPACE"
-
-      - name: Install dependencies
-        run: |
-          python -m pip install --upgrade pip
-          pip install -r pytorch/requirements.txt
-
-      - name: Check Formatting
-        run: |
-          black --check pytorch/
-
-      - name: Build frontends
-        run: |
-          pip install --no-build-isolation \
-            -Ccmake.build-type=Debug \
-            -v -e python/
-          pip install --no-build-isolation \
-            -v -e pytorch/
-
-      - name: Unit Tests
-        env:
-          DOCC_ACCESS_TOKEN: ${{ secrets.DOCC_CI_TOKEN }}
-        run: |
-          coverage run --source=pytorch/docc -m pytest -v pytorch/tests
-          pytest -v pytorch/tests --target=sequential
-          pytest -v pytorch/tests --target=openmp
-          coverage xml
-
-      - name: Coveralls
-        if: matrix.python-version == '3.14'
-        uses: coverallsapp/github-action@v2
-        with:
-          flag-name: pytorch
-          parallel: true
-          format: cobertura
-          file: coverage.xml
-
-  pytorch-tests-cuda:
-    runs-on:
-      group: dahlia
-      labels: cuda
-    container:
-      image: daisytuner/docc-run-env-llvm19-ubuntu-24.04:latest-amd64
-      options: >-
-        --cap-add=PERFMON
-        --gpus=all
-    strategy:
-      fail-fast: false
-      matrix:
-        python-version: ["3.11", "3.12", "3.13", "3.14"]
-
-    steps:
-      - uses: actions/checkout@v4
-        with:
-          submodules: recursive
-
-      - name: Set up Python ${{ matrix.python-version }}
-        uses: actions/setup-python@v5
-        with:
-          python-version: ${{ matrix.python-version }}
-
-      - name: Create virtual environment
-        run: |
-          python -m venv .venv
-          echo "$PWD/.venv/bin" >> $GITHUB_PATH
-
-      - name: Mark GitHub Actions workdir as safe
-        run: git config --global --add safe.directory "$GITHUB_WORKSPACE"
-
-      - name: Install dependencies
-        run: |
-          python -m pip install --upgrade pip
-          pip install -r pytorch/requirements-cuda.txt
-
-      - name: Check Formatting
-        run: |
-          black --check pytorch/
-
-      - name: Build frontends
-        run: |
-          pip install --no-build-isolation \
-            -Ccmake.build-type=Debug \
-            -v -e python/
-          pip install --no-build-isolation \
-            -v -e pytorch/
-
-      - name: Unit Tests
-        run: |
-          pytest -v pytorch/tests --target=cuda
-
-  pytorch-tests-rocm:
-    runs-on:
-      group: dahlia
-      labels: rocm
-    container:
-      image: daisytuner/docc-run-env-llvm19-ubuntu-24.04:latest-amd64
-      options: >-
-        --cap-add=PERFMON
-        --device /dev/kfd
-        --device /dev/dri
-    strategy:
-      fail-fast: false
-      matrix:
-        python-version: ["3.11", "3.12", "3.13"]
-
-    steps:
-      - uses: actions/checkout@v4
-        with:
-          submodules: recursive
-
-      - name: Set up Python ${{ matrix.python-version }}
-        uses: actions/setup-python@v5
-        with:
-          python-version: ${{ matrix.python-version }}
-
-      - name: Create virtual environment
-        run: |
-          python -m venv .venv
-          echo "$PWD/.venv/bin" >> $GITHUB_PATH
-
-      - name: Mark GitHub Actions workdir as safe
-        run: git config --global --add safe.directory "$GITHUB_WORKSPACE"
-
-      - name: Install dependencies
-        run: |
-          python -m pip install --upgrade pip
-          pip install -r pytorch/requirements-rocm.txt
-
-          # Build cupy-rocm from source for ROCm 6x
-          # To be removed when docker image is updated to ROCm 7x
-          export CUPY_INSTALL_USE_HIP=1
-          export ROCM_HOME=/opt/rocm
-          export HCC_AMDGPU_TARGET=$(rocminfo | grep -m1 -o 'gfx[0-9a-f]*')
-          pip install cupy --no-cache-dir
-
-      - name: Check Formatting
-        run: |
-          black --check pytorch/
-
-      - name: Build frontends
-        run: |
-          pip install --no-build-isolation \
-            -Ccmake.build-type=Debug \
-            -v -e python/
-          pip install --no-build-isolation \
-            -v -e pytorch/
-
-      - name: Unit Tests
-        run: |
-          pytest -v pytorch/tests --target=rocm
