+++ .github/workflows/unit_tests_linux.yml
@@ -475,6 +475,72 @@ jobs:
           format: cobertura
           file: coverage.xml
 
+  pytorch-tests-linux:
+    runs-on:
+      group: dahlia
+      labels: Linux
+    container:
+      image: daisytuner/docc-run-env-llvm19-ubuntu-24.04:latest-amd64
+      options: >-
+        --cap-add=PERFMON
+    strategy:
+      fail-fast: false
+      matrix:
+        python-version: ["3.11", "3.12", "3.13", "3.14"]
+
+    steps:
+      - uses: actions/checkout@v4
+        with:
+          submodules: recursive
+
+      - name: Set up Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v5
+        with:
+          python-version: ${{ matrix.python-version }}
+
+      - name: Create virtual environment
+        run: |
+          python -m venv .venv
+          echo "$PWD/.venv/bin" >> $GITHUB_PATH
+
+      - name: Mark GitHub Actions workdir as safe
+        run: git config --global --add safe.directory "$GITHUB_WORKSPACE"
+
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install -r pytorch/requirements.txt
+
+      - name: Check Formatting
+        run: |
+          black --check pytorch/
+
+      - name: Build frontends
+        run: |
+          pip install --no-build-isolation \
+            -Ccmake.build-type=Debug \
+            -v -e python/
+          pip install --no-build-isolation \
+            -v -e pytorch/
+
+      - name: Unit Tests
+        env:
+          DOCC_ACCESS_TOKEN: ${{ secrets.DOCC_CI_TOKEN }}
+        run: |
+          coverage run --source=pytorch/docc -m pytest -v pytorch/tests
+          coverage xml
+          pytest -v pytorch/tests --target=sequential
+          pytest -v pytorch/tests --target=openmp
+
+      - name: Coveralls
+        if: matrix.python-version == '3.14'
+        uses: coverallsapp/github-action@v2
+        with:
+          flag-name: pytorch
+          parallel: true
+          format: cobertura
+          file: coverage.xml
+
   ##### TARGETS #####
 
   cuda-tests-linux:
@@ -554,6 +620,15 @@ jobs:
 
           # Segformer correctness test
           pytest -v mlir/benchmarks/torch/model_zoo/segformer.py -k "cuda"
+
+      - name: PyTorch Tests
+        run: |
+          python -m venv .venv-pytorch
+          .venv-pytorch/bin/python -m pip install --upgrade pip
+          .venv-pytorch/bin/pip install -r pytorch/requirements-cuda.txt
+          export PYTHONPATH=$PWD/python:$PWD/pytorch
+
+          .venv-pytorch/bin/pytest -v pytorch/tests --target=cuda
 
   rocm-tests-linux:
     runs-on:
@@ -644,6 +719,22 @@ jobs:
 
           pytest -v mlir/benchmarks/torch/model_zoo/segformer.py -k "rocm"
 
+      - name: PyTorch Tests
+        run: |
+          python -m venv .venv-pytorch
+          .venv-pytorch/bin/python -m pip install --upgrade pip
+          .venv-pytorch/bin/pip install -r pytorch/requirements-rocm.txt
+          export PYTHONPATH=$PWD/python:$PWD/pytorch
+
+          # Build cupy-rocm from source for ROCm 6x
+          # To be removed when docker image is updated to ROCm 7x
+          export CUPY_INSTALL_USE_HIP=1
+          export ROCM_HOME=/opt/rocm
+          export HCC_AMDGPU_TARGET=$(rocminfo | grep -m1 -o 'gfx[0-9a-f]*')
+          .venv-pytorch/bin/pip install cupy --no-cache-dir
+
+          .venv-pytorch/bin/pytest -v pytorch/tests --target=rocm
+
   tenstorrent-tests-linux:
     runs-on:
       group: dahlia
@@ -711,7 +802,7 @@ jobs:
           pytest -v python/tests mlir/integration -m "tenstorrent"
 
   finish:
-    needs: [cpp-tests-linux, python-tests-linux, mlir-tests-linux, llvm-tests-linux]
+    needs: [cpp-tests-linux, python-tests-linux, mlir-tests-linux, llvm-tests-linux, pytorch-tests-linux]
     runs-on: ubuntu-latest
     steps:
       - name: Coveralls Finished
