+++ pytorch/docc/pytorch/graph_parser/pooling.py
@@ -1,3 +1,7 @@
+"""
+GraphParser modules for parsing pooling layers.
+"""
+
 import torch.fx
 from torch.fx.node import Argument
