+++ pytorch/docc/pytorch/graph_parser/reduction.py
@@ -1,3 +1,7 @@
+"""
+GraphParser modules for parsing reduction operations.
+"""
+
 import torch.fx
 from torch.fx.node import Argument
