+++ pytorch/requirements-base.txt
@@ -1,11 +1,12 @@
-pybind11
+numpy>=1.19.0
+ml_dtypes>=0.5.4
+sympy>=1.14.0
+scikit-build-core>=0.10
+pybind11>=2.12
+build
+scipy
+
 pytest
 coverage
 black==25.9.0
-build
-scikit-build-core
-sympy
-numpy
-ml_dtypes
-scipy
 requests
